refactor: reorganize modules
move legacy deserialization to own submodule
